backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-genai==1.0.0
anthropic>=0.40.0
python-dotenv==1.0.1
Pillow==11.1.0
pydantic==2.10.4
qrcode[pil]==8.0
requests==2.32.3
beautifulsoup4==4.12.3
